+++ controller/hybridgateway/route/status_test.go
@@ -1972,79 +1972,6 @@ func TestFilterOutGVKByKind(t *testing.T) {
 	}
 }
 
-func TestIsBackendRefSupported(t *testing.T) {
-	tests := []struct {
-		name  string
-		group *gwtypes.Group
-		kind  *gwtypes.Kind
-		want  bool
-	}{
-		{
-			name:  "nil group, Service kind",
-			group: nil,
-			kind:  kindPtr("Service"),
-			want:  true,
-		},
-		{
-			name:  "empty group, Service kind",
-			group: groupPtr(""),
-			kind:  kindPtr("Service"),
-			want:  true,
-		},
-		{
-			name:  "core group, Service kind",
-			group: groupPtr("core"),
-			kind:  kindPtr("Service"),
-			want:  true,
-		},
-		{
-			name:  "corev1 group, Service kind",
-			group: groupPtr("corev1"),
-			kind:  kindPtr("Service"),
-			want:  false,
-		},
-		{
-			name:  "v1 group, Service kind",
-			group: groupPtr("v1"),
-			kind:  kindPtr("Service"),
-			want:  false,
-		},
-		{
-			name:  "unsupported group, Service kind",
-			group: groupPtr("foo"),
-			kind:  kindPtr("Service"),
-			want:  false,
-		},
-		{
-			name:  "core group, unsupported kind",
-			group: groupPtr("core"),
-			kind:  kindPtr("Deployment"),
-			want:  false,
-		},
-		{
-			name:  "nil kind defaults to Service",
-			group: groupPtr("core"),
-			kind:  nil,
-			want:  true,
-		},
-		{
-			name:  "empty kind defaults to Service",
-			group: groupPtr("core"),
-			kind:  kindPtr(""),
-			want:  true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := IsBackendRefSupported(tt.group, tt.kind)
-			if got != tt.want {
-				t.Errorf("IsBackendRefSupported(%v, %v) = %v, want %v", tt.group, tt.kind, got, tt.want)
-			}
-		})
-	}
-}
-
 func TestIsRouteReferenceGranted(t *testing.T) {
 	tests := []struct {
 		name          string
