+++ controller/hybridgateway/upstream/policy.go
@@ -11,7 +11,6 @@ import (
 
 	configurationv1alpha1 "github.com/kong/kong-operator/v2/api/configuration/v1alpha1"
 	configurationv1beta1 "github.com/kong/kong-operator/v2/api/configuration/v1beta1"
-	"github.com/kong/kong-operator/v2/controller/hybridgateway/route"
 	"github.com/kong/kong-operator/v2/controller/hybridgateway/utils"
 	"github.com/kong/kong-operator/v2/controller/pkg/log"
 	gwtypes "github.com/kong/kong-operator/v2/internal/types"
@@ -46,7 +45,7 @@ func upstreamPolicyForRouteRule[R gwtypes.SupportedRouteRule](
 	}
 
 	for _, backendRef := range backendRefs {
-		if !route.IsBackendRefSupported(backendRef.Group, backendRef.Kind) {
+		if !utils.IsBackendRefSupported(backendRef.Group, backendRef.Kind) {
 			continue
 		}
