+++ controller/hybridgateway/utils/backendref_test.go
@@ -10,6 +10,79 @@ import (
 	gwtypes "github.com/kong/kong-operator/v2/internal/types"
 )
 
+func TestIsBackendRefSupported(t *testing.T) {
+	tests := []struct {
+		name  string
+		group *gwtypes.Group
+		kind  *gwtypes.Kind
+		want  bool
+	}{
+		{
+			name:  "nil group, Service kind",
+			group: nil,
+			kind:  new(gwtypes.Kind("Service")),
+			want:  true,
+		},
+		{
+			name:  "empty group, Service kind",
+			group: new(gwtypes.Group("")),
+			kind:  new(gwtypes.Kind("Service")),
+			want:  true,
+		},
+		{
+			name:  "core group, Service kind",
+			group: new(gwtypes.Group("core")),
+			kind:  new(gwtypes.Kind("Service")),
+			want:  true,
+		},
+		{
+			name:  "corev1 group, Service kind",
+			group: new(gwtypes.Group("corev1")),
+			kind:  new(gwtypes.Kind("Service")),
+			want:  false,
+		},
+		{
+			name:  "v1 group, Service kind",
+			group: new(gwtypes.Group("v1")),
+			kind:  new(gwtypes.Kind("Service")),
+			want:  false,
+		},
+		{
+			name:  "unsupported group, Service kind",
+			group: new(gwtypes.Group("foo")),
+			kind:  new(gwtypes.Kind("Service")),
+			want:  false,
+		},
+		{
+			name:  "core group, unsupported kind",
+			group: new(gwtypes.Group("core")),
+			kind:  new(gwtypes.Kind("Deployment")),
+			want:  false,
+		},
+		{
+			name:  "nil kind defaults to Service",
+			group: new(gwtypes.Group("core")),
+			kind:  nil,
+			want:  true,
+		},
+		{
+			name:  "empty kind defaults to Service",
+			group: new(gwtypes.Group("core")),
+			kind:  new(gwtypes.Kind("")),
+			want:  true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := IsBackendRefSupported(tt.group, tt.kind)
+			if got != tt.want {
+				t.Errorf("IsBackendRefSupported(%v, %v) = %v, want %v", tt.group, tt.kind, got, tt.want)
+			}
+		})
+	}
+}
+
 func TestHTTPBackendRefsToBackendRefs(t *testing.T) {
 	port80 := gatewayv1.PortNumber(80)
 	port443 := gatewayv1.PortNumber(443)
